// src/screens/main/profile/profileDetails/types.ts
export interface ProfileDetailsFormValues {
  firstName: string;
  lastName: string;
  username: string;
  email: string;
  countryCode: string;
  phoneNumber: string;
  bio: string;
  expertise: string[];
}

export interface ProfileDetailsData {
  firstName: string;
  lastName: string;
  username: string;
  email: string;
  countryCode: string;
  phoneNumber: string;
  bio: string;
  expertise: string[];
}